+++ CHANGELOG.rst
@@ -1,2 +1,8 @@
 Change Log
 ==========
+
+0.0.1 - Alpha
+-------------
+* Feature: Local File Storage
+* Feature: S3 File Storage
+* Feature: S3 Gevented File Storage